+++ comm/src/main/scala/coop/rchain/p2p/effects/Kvs.scala
@@ -1,7 +1,7 @@
-package coop.rchain.catscontrib
+package coop.rchain.p2p.effects
 
 import cats._, cats.data._, cats.implicits._
-import Catscontrib._
+import coop.rchain.catscontrib._, Catscontrib._
 
 trait Kvs[F[_], K, V] {
   def keys: F[Vector[K]]
