+++ comm/src/main/scala/coop/rchain/p2p/protocol.scala
@@ -5,6 +5,7 @@ import com.google.protobuf.ByteString
 import com.google.protobuf.any.{Any => AnyProto}
 import coop.rchain.comm.protocol.rchain._
 import coop.rchain.comm.protocol.routing
+import coop.rchain.p2p.effects._
 import cats._, cats.data._, cats.implicits._
 import coop.rchain.catscontrib._, Catscontrib._, Encryption.{Key, Nonce}
